LPS-123793 Forces the value of readOnly in the Field component and avoids traversing all fields
The implementations in Metal.js and the old codebase abuse Visitor a lot in the hot path, it is a very expensive operation to apply, so we took advantage of the component in the hierarchy to do this and avoid using Visitor excessively.

Some uses of PagesVisitor in the hot path will be removed as we remove Metal.js and allocate these changes/aggregation of data in the right places.