LPS-123793 Simplifies PageHeader implementation to handle localization and variation logic
Decouple the dependency of the EditablePageHeader component on the Page component and move the component to the MultiPagesVariant where it is only used.

Refactor the EditablePageHeader component to better handle localization, avoid using PageVisitor and use the index to change the title and description values according to the current language on the reducer.

Some of these changes are to simplify excess that deals with localization but also to remove the implementations that exist within the LayoutProvider that uses PageVisitor to change the title or description when the language changes.

The removal in LayoutProvider is not seen in this commit because these new changes will only reflect in Forms, the other products like Data Engine still use LayoutProvider and we will remove them in the next tickets as we migrate them.