LPS-123793 Add the Management Toolbar component to control the component rendered via JSP
This is a step to move the components to React and also remove the main responsibilities of the Forms, it is still necessary to move some other methods that control some buttons of the component but it will be done in another ticket.

Once we have all this well aligned, we can also start to migrate the components rendered in the JSP to React to facilitate the integration and composition of the application.