LPS-123793 Adds the NavigationBar component to handle the component rendered via JSP
This is another move to remove the responsibilities of the Forms main component, it also simplifies the logic to deal with routing, we will be using the React Router and other components to take advantage of the React lifecycle to make it visible and not visible depending on the route.

This component only interacts with the component rendered via JSP, makes it simpler to remove it from the JSP and bring it to React in the near future.