LPS-123793 Replaces the use of usePage in favor of useFormState in the fields
Well, usePage now only contains relevant information from the current page and no longer the state of the application, the responsibility now lies with useFormState.

Some small changes in the organization of the declaration of variables in the ReactFieldBase to improve readability, it is still necessary to make many changes in this component that is taken care of a lot of responsibility.