LPS-127583 Allowing rollback to happen
By catching the exceptions inside the doTransactional method we were
effectively making the transaction successful. As a result we were not
running the rollback and ended up persisting data (DDMStructureLayout)
that should have never been persisted.

I've decided to directly call TransactionInvokerUtil.invoke to avoid
refactoring the catch clauses. When using BaseTransactionalMVCActionCommand
exceptions are wrapped inside a PortletException and we would have to manually
check the type with instanceof which is uglier and less readable than the
nice catches that we have.