LPS-123793 Fixes error when changing the language back to recently typed text
Sidebar uses the focusedField information to use the settingsContext and render the fields, despite updating the data in the page structure in the builder, the focusedField was not updated and when it caused a rendering in the application it returned to the previous state that was in the focusedField.

The problem is that whenever you update the data in a field, you have to update the data in the focusedField and the page structure to avoid having old data problems, this can be a very silent problem, the focusedField must contain only the field id for reference but not all data, this requires more memory allocation and unnecessary duplication but for that we would have to change a lot of logic, leaving it as a suggestion for the team to change in the future.