LPS-124147 Adds the Report component to control the visibility of the Entries page
The same as we did for ManagementToolbar and NavigationBar, removing the responsibility of main.es.js and taking advantage of the React and React Router lifecycle to deal with this. The idea in the future is that we can actually import the component into the React structure without using the taglib.