LPS-124146 Rewrites AutoSave and StateSyncronizer in React.js
Well, this is a literal translation of the two components in React.js, few changes have been made to adapt to the React life cycle.

There are many DOM manipulations so that we can save the data in the DOM via `input[hidden]`, it may not be necessary anymore, a study in this area is needed to validate and remove these manipulations.