LPS-127832 Manage boolean dynamic attributes
In this change, we're changing the `buildDynamicAttributes` method of
the [InlineUtil](https://github.com/liferay/liferay-portal/blob/7985dd55f78d0b4e4960d1b8b07158787e32d29f/util-taglib/src/com/liferay/taglib/util/InlineUtil.java) class to handle boolean attributes.

According to the [spec](https://html.spec.whatwg.org/#boolean-attribute), the presence of
a boolean attribute on an element represents the "true" value, and the
absence of the attribute represents the "false" value, but the values
"true" and "false" are not allowed on boolean attributes. To represent a
false value, the attribute has to be omitted altogether.

If a dynamic attribute is one of the 25 available boolean attributes (see [here](https://meiert.com/en/blog/boolean-attributes-of-html/))
and it's value is "false", we omit writting it and if it's value is
"true" we replace it with the name of the attribute.

**Test plan**
- Fetch this branch
- Deploy the `util-taglib` module
- Add this code to a jsp, for example [buttons.jsp](https://github.com/liferay/liferay-portal/blob/85f66220f377144c4cf025be77a04c763a9f04b2/modules/apps/frontend-taglib/frontend-taglib-clay-sample-web/src/main/resources/META-INF/resources/partials/buttons.jsp)

```jsp
<clay:button disabled="<%= true %>" label="Disabled" />
<clay:button disabled="<%= false %>" label="Enabled" />
```

- Assert that it renders
    - A button with the `disabled` attribute (and no value)
    - A button without a `disabled` attribute