LPS-127671 Revert "Set minimum column size two 2 and adjust resizing"
This reverts commit 247cbe14bcde321b3a8f287585e76502139478cb.

We are reversing this implementation because it implements a behavior different from what would be expected, we must consider that the minimum column size must be only 2 but allow the resize to be done one by one. The way it was implemented, the user is unable to create a column with size 3, only two by two, that is, just even numbers

An important point is that we should test if the Form loads a form with non-even fields, the resize will correct and mess up the layout since it will only have even-sized columns.