LPS-128262 Replace `runLegacyBuild` with the new `mode` setting
We needed to introduce a new setting in npm-scripts to distinguish when to run
webpack only, webpack + bundler, or only bundler.

This is because now we are inheriting federation config for every project in
liferay-portal, which makes it impossible to decide what needs to be done based
on the existence/absence of federation configuration.

Regarding `runLegacyBuild`, it can be removed because it is redundant (we can
infer everything from the `mode` flag alone)