LPS-127081 Add frontend-js-state-web
This is an squashed version of the draft previously published here:

    https://github.com/wincent/liferay-portal/pull/373

with some various bits added.

Doc comments in code explain what this does and design trade-offs are
explained in depth in this PR:

    https://github.com/liferay-frontend/liferay-portal/pull/874

but the TL;DR is that this is a type-safe mechanism for sharing global
state across different apps inside liferay-portal, with an emphasis on
having a minimal API footprint. We have several options for growing the
API in the future (discussed on the PR), but I want to start small.