LPS-128506 Adds necessary ClayTooltipProvider to MultiSelect component
In this case, this is necessary because of how the component is passing
`otherProps` to a specific element within the element hierarchy and
not directly to a container. This makes the mouse listeners to be added
to the wrong element and not triggered in other parts of the component.

To solve this in this case we can simply add one additional TooltipProvider
which targets an ancestor wrapper and can handle the mouse interactions
properly