LPS-127081 Fix lint errors
Some of these are automatic (eg. import ordering), and some of them were
manual (eg. object key sorting).

Note that I had to add an .eslintrc.js file to reconcile a disagreement
about "use before define" between TS and ESLint. Type definitions can be
recursive and circular, so we need to turn off the latter rule and let
the TS-specific rule handle that for us.