LPS-129340 Serialize the form builder context in-depth to avoid JSONArray and JSONObject in the structure
It seems that just getting `getContext` does not bring the context fully serialized and when we try to create a new `JSONObject` from it, some properties of some Fields that use `jsonFactory` end up being added as part of the structure, as an example `rows: {JSONArray: []}` instead of `rows: []`.

The solution here was to serialize the context in-depth and then create a JSONObject, this seems to have removed the use cases known above.