LPS-129182 Fix React fragments
We had to rename the `frontend-js-react-web` project to
`@liferay/frontend-js-react-web` because of a security vulnerability. That made
React fragments fail because they were referring to the old location.

After discussing it with the Echo team, we decided to replace occurrences of
`'frontend-js-react-web$react'` in the fragments' source code by the new name.
This is to avoid DB upgrade processes and/or fragment projects recompilation.

Also, because `'frontend-js-react-web$react'` as a placeholder is a bit ugly, we
will now support `__PORTAL_REACT__` as the canonical form and will change the
toolkit accordingly.

In any case, we need to leave `'frontend-js-react-web$react'` for backward
compatibility (and also as a reminder of how not to do it in the future ;-)).