LPS-129346 Prevents moving a FieldGroup into itself and into another FieldGroup
The LayoutProvider implementation did not deal well with preventing a FieldGroup from moving inside into itself, because it did not give Feedback to the user that this is not possible and was an expensive operation when trying to search for the target within the source with the visitor and also it is not the appropriate place to be added, it is the responsibility of the DnD mechanism.

There were also other untreated behaviors, such as:

- Drop the common Field inside itself
- Drop a FieldGroup inside any Field (This means that the FieldGroup can be added in a placeholder.)

These behaviors were not perceived because there was a `try catch` that captured the handlers' errors and suppressed what is invisible to the user if it worked or was wrong and it was also an insecure and sensitive way to deal with these use cases.