LPS-129801 Avoid use of ambient types in frontend-js-state-web
My experimentation with project references (next commit) and manually
wiring things up (previous commit) shows that ambient types probably
aren't going to work so well in a cross-project scenario. TypeScript
doesn't pick them up automatically, except inside the project where they
are defined.

So, this is a refactoring commit that just changes this file into a
"real" module (that doesn't export any values, but does export types),
and we import then types explicitly wherever we need them. The module
should be elided from the actual output and never imported at runtime
because nobody is importing values from it (the type imports get
elided). As you can see here, Babel just makes an empty module for it
that nobody is ever going to use (wrapped here for readability; in
reality the file is just a single line):

    Liferay.Loader.define(
        "@liferay/frontend-js-state-web@1.0.0/types",
        ["module", "exports", "require"],
        function (e, r, i) {
            window;
        }
    );