LPS-129367 Remove exclusion customization to avoid build failure
This exclusion is configured as `lodash: true` in the base config.

npm-scripts doesn't know how to merge a boolean and an array. We could modify them
to cover this case, but if we can simply avoid the case it is much better, because
merging booleans and arrays, even though we may give it some semantics, sounds a
bit weird.

So I'm removing the exclude to see if anything bad happens. Manually testing
everything seems OK. Also, inspecting the code I don't see where lodash is used (in
segments) so I think this may be a leftover.

Let's see what the CI has to say...