LPS-129899 Reassemble the components of the variant in the Layout respecting the editable property in the nesting level
The Layout component is used nested to render elements of the FieldGroup or FieldSet. FieldSet component uses Layout to render nested fields but disables the editable option to disable any features for elements within the FieldSet.

The strategy here is simple, to avoid analyzing all fields or nested fields as was being done in a hot path that can generate performance problems and delay the rendering of the elements, we just revalidated the merge of the variants within the Layout component. Because once the editable is false for FieldSet the component will only use the default variants, without DnD, Resizable...

Try to avoid using visitors as much as possible for these hot path cases.