LPS-129529 Adds the Data representation implementation of Forms to Data Engine
The `useDataView` hook is the component that ends up doing the entire orchestration of creating the representation of the data, reusing, revalidating, caching, and returning the output of the representation.

Some resources make it very valuable and significant for the growth of the application and how it makes it easier for two different structures to coexist but with the same raw data.

- The components never know that these two different structures exist, because there is no need to call any method to convert the data.
- It is easy to slowly migrate components to use the new data structure.
- It will be simpler to maintain data representations based on the implementation of "schema" compared to data conversion.
- We don't couple any method or class to the components that we need to call for every corner, like DataLayoutBuilder that was done before.
- It is not necessary to wait to convert all the data before starting to render the component or to call any other method.
- We create data representation whenever possible using reference to avoid overloading the creation of new objects being allocated in the heap.
- When a Schema is created for the first time it is reused so we don't need to create another one again, there is a reuse of schema in different instances of the components if you use the same representation.
- There is a revalidation of the raw data to avoid loss of reference.

Despite having a lot of good things with this implementation I would say that there is no contraindication just that it adds minimal overhead to the data output which is incredibly small in the 100ns range but that this is a great pros due to the previous implementation that did a lot data conversion all the time which was considerably noticeable at times the slowness.

The `useDataView` hook is not used directly, it is built into `useFormState` to avoid coupling its use to the components. The logic is not done on top of `useFormState` because it is based on principles of responsibility, this helps when all components are using the Data Engine data structure, the swap will be easy to do.

This implementation is not yet complete, it does not cover the case of creating a reducer using the state represented in the Data Engine structure. Adding in next interactions.

Also, not all data representations were added, just adding the representation for DataLayout and DataDefinition initially. Add more methods or more Schemas as needed.